{{cookiecutter.project_slug}}/requirements/base.txt
Django==2.1.5
psycopg2-binary==2.7.6.1
dj-database-url==0.5.0
whitenoise==4.1.2
channels==2.1.6
channels_redis==2.3.3
service_identity==18.1.0
django-allauth==0.38.0
werkzeug==0.14.1
pyopenssl==18.0.0
django-js-reverse==0.8.2
djangorestframework==3.9.0
django-redis==4.10.0
django-log-request-id==1.3.2
raven==6.10.0
rq==0.12  # pyup: <0.13
django-rq==1.2.0  # pyup: <1.3.0
django-rq-scheduler==1.1.1
rq-scheduler==0.8.3  # pyup: ==0.8.3
django-storages==1.7.1{% if cookiecutter.use_bucketeer_aws_for_file_storage == 'y' %}
boto3==1.9.78{% endif %}
django-hashid-field==2.1.5
redis==2.10.6  # pyup: <3.0
logfmt==0.4
